reorg of python requires tests
